Fixed issue involving display of relationship references.
The generated scaffold now contains code to replace a JSON object reference in an AngularJS Resource, with a reference to a Resource object present in a Resource collection. This is done to pass the object identity (===) comparison test used by AngularJS for HTML5 select elements. AngularJS requires that the ngModel bound to a select, should also be present in the ngOptions array/collection, failing which it does not display the bound model element. When a Resource with a JSON object is originally obtained, the JSON object is not equal (===) to any element in the Resource array/collection used by ngOptions; hence the need to replace the JSON object with an actual element from the array/collection.